refactor(memory): split dedupe.rs and entity_id_migration.rs into focused modules
Both files grew past the repo's file-size convention across several rounds of
adversarial test hardening, with #[cfg(test)] mod tests holding the bulk of
each file's lines. Split each into a directory (matching the existing
config/mod.rs and storage/memory/tests.rs precedent of a mod.rs plus sibling
files) with tests separated from implementation and grouped by concern:

- dedupe.rs (2181 lines) -> dedupe/{mod,test_support,tests,superseded_by_tests}.rs
  - mod.rs: DedupeSummary + dedupe_entity_ids/collapse_group_survivor/
    rewrite_cross_references (unchanged)
  - tests.rs: core merge/dry-run/rollback/basic-edge behavior
  - superseded_by_tests.rs: intra- and cross-group superseded_by
    reference-resolution edge cases (the adversarial-hardening rounds)
- entity_id_migration.rs (1199 lines) ->
  entity_id_migration/{mod,test_support,migration_tests,collision_recovery_tests}.rs
  - mod.rs: backfill_entity_ids/promote_entity_id_unique_index (unchanged)
  - migration_tests.rs: Step A/B population and index-promotion behavior
  - collision_recovery_tests.rs: add_note/add_note_superseding
    insert-then-recover behavior once the index is promoted

Pure refactor: every test relocated verbatim (33 dedupe tests, 22
entity_id_migration tests, exact parity with pre-split counts), no logic
changed. Full workspace suite green before and after, both
--features rich-formats and --no-default-features; cargo fmt and
cargo clippy --workspace --all-targets -- -D warnings clean on both
configs.
